Step 1/13 — promote shared contracts to @derecksnotes/shared.
Collapses six duplicated declarations and one drifted constant into
typed, single-source modules. Every later step imports these and stops
redeclaring shapes inline.

shared/src/types
- new auth.ts: UpdateProfileInput, AuthError (moved from user.ts).
- new admin.ts: AdminUser, AdminPendingComment, AdminReport,
  DashboardStats, DashboardResponse, AuditLogEntry, NotificationStats,
  AnalyticsData — all admin domain types now live next to each other.
- new comments.ts: CommentData, CommentsListResponse, RepliesResponse,
  CommentHistoryEntry, UserComment.
- new notifications.ts: NotificationType, NotificationEntry.
- new posts.ts: PostBasic, PostReactionResponse (PostStats kept as
  alias for the legacy /posts/stats endpoint), ReadHistoryEntry.
- api.ts trimmed to the generic envelopes only (PaginatedResponse,
  ApiSuccess, ApiError, SessionInfo).
- types/index.ts re-exports from the new domain files.

shared/src/constants
- new password.ts: PASSWORD_RULES + isPasswordValid() — used by client
  PasswordStrength and ready to be imported by the server zod schemas
  in step 5/7.
- new reports.ts: REPORT_REASONS tuple + ReportReason union +
  isValidReportReason guard — replaces the literal-string array
  duplicated between the client report dialog and the server zod enum.

Consumers
- client/src/context/AuthContext.tsx — updateProfile signature
  collapses to UpdateProfileInput.
- client/src/components/posts/PostReactionButtons.tsx — local Stats
  interface deleted; uses PostReactionResponse.
- client/src/app/admin/page.tsx — inline AdminReport interface
  deleted; imports from shared.
- client/src/app/account/page.tsx — local isPasswordValid +
  hard-coded "8+ characters" deleted; uses PASSWORD_RULES +
  sharedIsPasswordValid.

Tested: bunx tsc --noEmit clean on server + client; bun test 40 pass.
