+++ shared/src/types/index.ts
@@ -1,3 +1,8 @@
 export * from './api';
 export * from './user';
+export * from './auth';
+export * from './admin';
+export * from './comments';
+export * from './notifications';
+export * from './posts';
 export * from './search';
